+++ vector/src/main/java/im/vector/riotx/features/home/room/detail/readreceipts/DisplayReadReceiptItem.kt
@@ -33,6 +33,7 @@ abstract class DisplayReadReceiptItem : EpoxyModelWithHolder<DisplayReadReceiptI
     @EpoxyAttribute lateinit var matrixItem: MatrixItem
     @EpoxyAttribute var timestamp: CharSequence? = null
     @EpoxyAttribute lateinit var avatarRenderer: AvatarRenderer
+    @EpoxyAttribute var userClicked: (() -> Unit)? = null
 
     override fun bind(holder: Holder) {
         avatarRenderer.render(matrixItem, holder.avatarView)
@@ -43,6 +44,7 @@ abstract class DisplayReadReceiptItem : EpoxyModelWithHolder<DisplayReadReceiptI
         } ?: run {
             holder.timestampView.isVisible = false
         }
+        holder.view.setOnClickListener { userClicked?.invoke() }
     }
 
     class Holder : VectorEpoxyHolder() {